central_data_repository_application/requirements.txt
Flask
Werkzeug
psycopg2-binary
pandas
